pyproject.toml
[project]
name = "django-ca-cmc"
version = "0.1.0"
description = "django-ca plugin for Certificate Management over CMS (CMC)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ca>=2.1.1",
    "python-cmc>=1.22",
]

[dependency-groups]
dev = [
    "django-stubs>=5.1.2",
    "mypy>=1.14.1",
    "pytest-django>=4.9.0",
    "pytest>=8.3.4",
    "pytest-random-order>=1.1.1",
    "ruff>=0.9.2",
    "coverage[toml]>=7.6.10",
    "pytest-cov>=6.0.0",
]

[tool.coverage.run]
source = [
    "django_ca_cmc",
]
branch = true

[tool.django-stubs]
django_settings_module = "demo.settings"

[tool.mypy]
strict = true
show_error_codes = true
mypy_path = "django_ca_cmc:demo"
plugins = ["mypy_django_plugin.main"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = [
    "--cov",
    "--cov-report=html:coverage/",
    "--cov-fail-under=100",
    "--random-order",
]
pythonpath = [
    "demo",
    ".",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = [
    "B",  # flake8-bugbear
    "D",  # pydocstyle
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "G",  # flake8-logging-format
    "DJ",  # flake8-django
    "PL",  # pylint
    "PT",  # flake8-pytest-style
    "UP",  # pyupgrade
    "RUF",  # ruff specific checks
]

ignore = [
    "D105",
    "D203",
    "D212",
]
